WIP: UTIL_EST: use estimated utilization on energy aware wakeup path
The energy aware wakeup path is in charge to select the most suitable
CPU to scheduler a task. The usage of an estimated utilization for the
task to allocate, as well as for the evaluation of the capacity available
on candidate CPUs, can provide a more stable information to refine the
scheduling decisions.

This patch updates all the energy aware wakeup paths to use the estimated
utilization whenever possible, i.e. when the UTIL_EST feature is enabled.
Specifically we use the estimated utilization to support:
a) CPU selection
b) energy module computation,
   i.e. for the computation of the (estimated) active and idle periods

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: geckyn <geckyn@gmail.com>